extract data parser helper
